fix custom chunk
