+++ requirements.txt
@@ -6,6 +6,7 @@ idna==2.6
 macholib==1.9
 pefile==2017.11.5
 protobuf==3.5.1
+pypng==0.0.18
 requests==2.18.4
 six==1.11.0
 urllib3==1.22
